Renamed playerctl.rs to media_player.rs
